+++ lib/another_tus_client.dart
@@ -3,3 +3,4 @@ export 'src/exceptions.dart';
 export 'src/store.dart';
 export 'src/retry_scale.dart';
 export 'src/file_utils.dart';
+export 'src/upload_manager.dart';